requirements.txt
vk_api
python-dotenv
openai
schedule